FEATURE: revert disallowing putting URLs in titles for TL0 users
This reverts a part of changes introduced by
https://github.com/discourse/discourse/pull/13947

In that PR I:
1. Disallowed topic feature links for TL-0 users
2. Additionally, disallowed just putting any URL in topic titles for TL-0 users

Actually, we don't need the second part. It introduced unnecessary
complexity for no good reason. In fact, it tries to do the job that
anti-spam plugins (like Akismet plugin) should be doing.

This PR reverts this second change.